Build info refactoring
