+++ public/app/features/alerting/unified/RuleEditor.test.tsx
@@ -20,7 +20,7 @@ import { ExpressionEditorProps } from './components/rule-editor/ExpressionEditor
 import { searchFolders } from '../../../../app/features/manage-dashboards/state/actions';
 import * as api from 'app/features/manage-dashboards/state/actions';
 import { GrafanaAlertStateDecision, PromApplication } from 'app/types/unified-alerting-dto';
-import { fetchBuildInfo } from './api/prometheus';
+import { fetchBuildInfo } from './api/buildInfo';
 
 jest.mock('./components/rule-editor/ExpressionEditor', () => ({
   // eslint-disable-next-line react/display-name
@@ -29,7 +29,7 @@ jest.mock('./components/rule-editor/ExpressionEditor', () => ({
   ),
 }));
 
-jest.mock('./api/prometheus');
+jest.mock('./api/buildInfo');
 jest.mock('./api/ruler');
 jest.mock('./utils/config');
 jest.mock('../../../../app/features/manage-dashboards/state/actions');
